+++ icekit/content_collections/page_type_plugins.py
@@ -1,0 +1,40 @@
+from django.conf.urls import patterns, url
+from django.core.exceptions import ObjectDoesNotExist
+from django.http import Http404
+from django.template.response import TemplateResponse
+from fluent_pages.extensions import page_type_pool
+from icekit.page_types.layout_page.admin import LayoutPageAdmin
+
+from icekit.plugins import ICEkitFluentContentsPagePlugin
+
+
+class ListingPagePlugin(ICEkitFluentContentsPagePlugin):
+    # render_template = 'icekit_content_collections/layouts/collection.html'
+    model_admin = LayoutPageAdmin
+
+    def get_view_response(self, request, page, view_func, view_args, view_kwargs):
+        """
+        Render the custom view that was exposed by the extra plugin URL patterns.
+        This gives the ability to add extra middleware logic.
+        """
+        return view_func(request, page, *view_args, **view_kwargs)
+
+    def collected_content_view(request, parent, slug):
+        try:
+            # using .visible() here to acknowledge IS_DRAFT context.
+            page = parent.get_visible_items().get(slug=slug)
+        except ObjectDoesNotExist:
+            raise Http404
+
+        # If the article defines its own response, use that.
+        if hasattr(page, 'get_response'):
+            return page.get_response(request, parent=parent)
+        else:
+            raise AttributeError("You need to define `%s.get_response(request, parent)`, or override `collected_content_view` in your `ListingPagePlugin` class" % (type(page).__name__))
+
+    urls = patterns('',
+        url(
+            '^(?P<slug>[-\w]+)/$',
+            collected_content_view,
+        ),
+    )