+++ icekit/project/settings/_production.py
@@ -10,23 +10,7 @@ CACHES['default'].update({
     'LOCATION': 'redis://redis:6379/1',
 })
 
-# EMAIL_HOST = ''
-# EMAIL_HOST_USER = ''
-
 LOGGING['handlers']['logfile']['backupCount'] = 100
-
-MIDDLEWARE_CLASSES = (
-    ('django.middleware.cache.UpdateCacheMiddleware', ) +
-    MIDDLEWARE_CLASSES +
-    ('django.middleware.cache.FetchFromCacheMiddleware', )
-)
-
-TEMPLATES_DJANGO['OPTIONS']['loaders'] = [
-    (
-        'django.template.loaders.cached.Loader',
-        TEMPLATES_DJANGO['OPTIONS']['loaders'],
-    ),
-]
 
 # CELERY EMAIL ################################################################
 
@@ -34,6 +18,4 @@ CELERY_EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
 
 # STORAGES ####################################################################
 
-# AWS_ACCESS_KEY_ID = ''
-AWS_STORAGE_BUCKET_NAME = SITE_SLUG
 ENABLE_S3_MEDIA = True