+++ icekit/project/settings/calculated.py
@@ -1,14 +1,14 @@
+from __future__ import absolute_import
+
 import importlib
 import os
 import sys
 
 from icekit.utils.sequences import dedupe_and_sort
 
-BASE_SETTINGS_MODULE = os.environ.setdefault('BASE_SETTINGS_MODULE', 'base')
-
-print '# BASE_SETTINGS_MODULE: %s' % BASE_SETTINGS_MODULE
-
 # Emulate `from ... import *` with base settings module from environment.
+BASE_SETTINGS_MODULE = os.environ.setdefault('BASE_SETTINGS_MODULE', 'base')
+print '# BASE_SETTINGS_MODULE: %s' % BASE_SETTINGS_MODULE
 try:
     locals().update(importlib.import_module(
         'icekit.project.settings._%s' % BASE_SETTINGS_MODULE).__dict__)
@@ -56,8 +56,6 @@ INSTALLED_APPS = dedupe_and_sort(
         'flat',
         'test_without_migrations',
     ],
-     # These apps last:
-     []
 )
 
 # Sort middleware according to documentation.