+++ icekit/project/settings/glamkit.py
@@ -1,0 +1,5 @@
+from .icekit import *
+
+# DJANGO ######################################################################
+
+INSTALLED_APPS += ('sponsors', )