+++ icekit/publishing/admin.py
@@ -240,13 +240,20 @@ class _PublishingHelpersMixin(object):
         if hasattr(obj, 'get_real_instance'):
             obj = obj.get_real_instance()
 
+        try:
+            published_obj_url = obj.get_absolute_url()
+            draft_obj_url = published_obj_url + '?edit'
+        except:
+            published_obj_url = draft_obj_url = None
+
         template_name = 'admin/publishing/_change_list_publishing_column.html'
         t = loader.get_template(template_name)
         c = Context({
             'object': obj,
             'has_publish_permission':
                 self.has_publish_permission(self.request, obj),
-            'img_path': settings.STATIC_URL + 'admin/img/',
+            'published_url': published_obj_url,
+            'draft_url': draft_obj_url,
         })
         try:
             if isinstance(obj, PublishingModel):
@@ -266,7 +273,7 @@ class PublishingAdmin(ModelAdmin, _PublishingHelpersMixin):
     # publish or unpublish actions sometime makes the plugins disappear from
     # page so we disable it for now, until we can investigate it further.
     # actions = (make_published, make_unpublished, )
-    list_display = ('publishing_object_title', 'publishing_column',)
+    list_display = ('publishing_object_title', 'publishing_column', 'publishing_modified_at')
     list_filter = (PublishingStatusFilter, PublishingPublishedFilter)
 
     actions = ['publish', 'unpublish']